Determining the Best Regression Model Across Orders and Parameters
We consider seven possible regression models. Our goal is to decide which one of the seven models is the "best" one to explain the observed data. For each model with specific parameters Beta, "goodness of fit" can be measured by the likelihood term P(y|Beta,M). Here, Beta is a vector of regression coefficients, and M indicates a polynomial regression model of order p. Model evidence is evaluated using P(y|M), which describes how likely the data are generated by a polynomial model.